+++ src/game/mod.rs
@@ -250,9 +250,14 @@ impl Game {
                 Ok(())
             }
             Err(character::Dead) => {
-                // leave hero items in the location
-                let tombstone = Chest::drop(self);
-                self.tombstones.insert(self.location.to_string(), tombstone);
+                // Drop hero items in the location. If there was a previous tombstone
+                // merge the contents of both chests
+                let mut tombstone = Chest::drop(self);
+                let location = self.location.to_string();
+                if let Some(previous) = self.tombstones.remove(&location) {
+                    tombstone.extend(previous);
+                }
+                self.tombstones.insert(location, tombstone);
 
                 Event::emit(self, Event::BattleLost);
                 Err(character::Dead)